Moving AbstractObjectEditor test to right place
